Spaghetti
An attempt to shortcut the Image stitching process. Didn't work, so I'm going to restart properly
